Changes following PR review comments
