+++ tests/system_tests/conftest.py
@@ -1,3 +1,5 @@
+import re
+from decimal import Decimal
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -5,6 +7,116 @@ from dodal.beamlines import i03
 from dodal.devices.oav.oav_parameters import OAVConfigParams
 from ophyd_async.core import AsyncStatus, set_mock_value
 from requests import Response
+
+# Map all the case-sensitive column names from their normalised versions
+DATA_COLLECTION_COLUMN_MAP = {
+    s.lower(): s
+    for s in [
+        "dataCollectionId",
+        "BLSAMPLEID",
+        "SESSIONID",
+        "experimenttype",
+        "dataCollectionNumber",
+        "startTime",
+        "endTime",
+        "runStatus",
+        "axisStart",
+        "axisEnd",
+        "axisRange",
+        "overlap",
+        "numberOfImages",
+        "startImageNumber",
+        "numberOfPasses",
+        "exposureTime",
+        "imageDirectory",
+        "imagePrefix",
+        "imageSuffix",
+        "imageContainerSubPath",
+        "fileTemplate",
+        "wavelength",
+        "resolution",
+        "detectorDistance",
+        "xBeam",
+        "yBeam",
+        "comments",
+        "printableForReport",
+        "CRYSTALCLASS",
+        "slitGapVertical",
+        "slitGapHorizontal",
+        "transmission",
+        "synchrotronMode",
+        "xtalSnapshotFullPath1",
+        "xtalSnapshotFullPath2",
+        "xtalSnapshotFullPath3",
+        "xtalSnapshotFullPath4",
+        "rotationAxis",
+        "phiStart",
+        "kappaStart",
+        "omegaStart",
+        "chiStart",
+        "resolutionAtCorner",
+        "detector2Theta",
+        "DETECTORMODE",
+        "undulatorGap1",
+        "undulatorGap2",
+        "undulatorGap3",
+        "beamSizeAtSampleX",
+        "beamSizeAtSampleY",
+        "centeringMethod",
+        "averageTemperature",
+        "ACTUALSAMPLEBARCODE",
+        "ACTUALSAMPLESLOTINCONTAINER",
+        "ACTUALCONTAINERBARCODE",
+        "ACTUALCONTAINERSLOTINSC",
+        "actualCenteringPosition",
+        "beamShape",
+        "dataCollectionGroupId",
+        "POSITIONID",
+        "detectorId",
+        "FOCALSPOTSIZEATSAMPLEX",
+        "POLARISATION",
+        "FOCALSPOTSIZEATSAMPLEY",
+        "APERTUREID",
+        "screeningOrigId",
+        "flux",
+        "strategySubWedgeOrigId",
+        "blSubSampleId",
+        "processedDataFile",
+        "datFullPath",
+        "magnification",
+        "totalAbsorbedDose",
+        "binning",
+        "particleDiameter",
+        "boxSize",
+        "minResolution",
+        "minDefocus",
+        "maxDefocus",
+        "defocusStepSize",
+        "amountAstigmatism",
+        "extractSize",
+        "bgRadius",
+        "voltage",
+        "objAperture",
+        "c1aperture",
+        "c2aperture",
+        "c3aperture",
+        "c1lens",
+        "c2lens",
+        "c3lens",
+        "startPositionId",
+        "endPositionId",
+        "flux",
+        "bestWilsonPlotPath",
+        "totalExposedDose",
+        "nominalMagnification",
+        "nominalDefocus",
+        "imageSizeX",
+        "imageSizeY",
+        "pixelSizeOnImage",
+        "phasePlate",
+        "dataCollectionPlanId",
+    ]
+}
 
 
 @pytest.fixture
@@ -82,3 +194,33 @@ def oav_for_system_test(test_config_files):
         ):
             oav.parameters.load_microns_per_pixel(1.0, 1024, 768)
             yield oav
+
+
+def compare_actual_and_expected(
+    id, expected_values, fetch_datacollection_attribute, column_map: dict | None = None
+):
+    results = "\n"
+    for k, v in expected_values.items():
+        actual = fetch_datacollection_attribute(
+            id, column_map[k.lower()] if column_map else k
+        )
+        if isinstance(actual, Decimal):
+            actual = float(actual)
+        if isinstance(v, float):
+            actual_v = actual == pytest.approx(v)
+        else:
+            actual_v = actual == v
+        if not actual_v:
+            results += f"expected {k} {v} == {actual}\n"
+    assert results == "\n", results
+
+
+def compare_comment(
+    fetch_datacollection_attribute, data_collection_id, expected_comment
+):
+    actual_comment = fetch_datacollection_attribute(
+        data_collection_id, DATA_COLLECTION_COLUMN_MAP["comments"]
+    )
+    match = re.search(" Zocalo processing took", actual_comment)
+    truncated_comment = actual_comment[: match.start()] if match else actual_comment
+    assert truncated_comment == expected_comment
