+++ tests/system_tests/hyperion/external_interaction/test_ispyb_dev_connection.py
@@ -1,10 +1,8 @@
 from __future__ import annotations
 
 import os
-import re
 from collections.abc import Callable, Sequence
 from copy import deepcopy
-from decimal import Decimal
 from typing import Any, Literal
 
 import pytest
@@ -52,6 +50,11 @@ from mx_bluesky.hyperion.parameters.gridscan import (
 )
 from mx_bluesky.hyperion.parameters.rotation import RotationScan
 
+from ...conftest import (
+    DATA_COLLECTION_COLUMN_MAP,
+    compare_actual_and_expected,
+    compare_comment,
+)
 from .conftest import raw_params_from_file
 
 EXPECTED_DATACOLLECTION_FOR_ROTATION = {
@@ -63,116 +66,6 @@ EXPECTED_DATACOLLECTION_FOR_ROTATION = {
     "synchrotronMode": SynchrotronMode.USER.value,
     "slitGapHorizontal": 0.123,
     "slitGapVertical": 0.234,
-}
-
-# Map all the case-sensitive column names from their normalised versions
-DATA_COLLECTION_COLUMN_MAP = {
-    s.lower(): s
-    for s in [
-        "dataCollectionId",
-        "BLSAMPLEID",
-        "SESSIONID",
-        "experimenttype",
-        "dataCollectionNumber",
-        "startTime",
-        "endTime",
-        "runStatus",
-        "axisStart",
-        "axisEnd",
-        "axisRange",
-        "overlap",
-        "numberOfImages",
-        "startImageNumber",
-        "numberOfPasses",
-        "exposureTime",
-        "imageDirectory",
-        "imagePrefix",
-        "imageSuffix",
-        "imageContainerSubPath",
-        "fileTemplate",
-        "wavelength",
-        "resolution",
-        "detectorDistance",
-        "xBeam",
-        "yBeam",
-        "comments",
-        "printableForReport",
-        "CRYSTALCLASS",
-        "slitGapVertical",
-        "slitGapHorizontal",
-        "transmission",
-        "synchrotronMode",
-        "xtalSnapshotFullPath1",
-        "xtalSnapshotFullPath2",
-        "xtalSnapshotFullPath3",
-        "xtalSnapshotFullPath4",
-        "rotationAxis",
-        "phiStart",
-        "kappaStart",
-        "omegaStart",
-        "chiStart",
-        "resolutionAtCorner",
-        "detector2Theta",
-        "DETECTORMODE",
-        "undulatorGap1",
-        "undulatorGap2",
-        "undulatorGap3",
-        "beamSizeAtSampleX",
-        "beamSizeAtSampleY",
-        "centeringMethod",
-        "averageTemperature",
-        "ACTUALSAMPLEBARCODE",
-        "ACTUALSAMPLESLOTINCONTAINER",
-        "ACTUALCONTAINERBARCODE",
-        "ACTUALCONTAINERSLOTINSC",
-        "actualCenteringPosition",
-        "beamShape",
-        "dataCollectionGroupId",
-        "POSITIONID",
-        "detectorId",
-        "FOCALSPOTSIZEATSAMPLEX",
-        "POLARISATION",
-        "FOCALSPOTSIZEATSAMPLEY",
-        "APERTUREID",
-        "screeningOrigId",
-        "flux",
-        "strategySubWedgeOrigId",
-        "blSubSampleId",
-        "processedDataFile",
-        "datFullPath",
-        "magnification",
-        "totalAbsorbedDose",
-        "binning",
-        "particleDiameter",
-        "boxSize",
-        "minResolution",
-        "minDefocus",
-        "maxDefocus",
-        "defocusStepSize",
-        "amountAstigmatism",
-        "extractSize",
-        "bgRadius",
-        "voltage",
-        "objAperture",
-        "c1aperture",
-        "c2aperture",
-        "c3aperture",
-        "c1lens",
-        "c2lens",
-        "c3lens",
-        "startPositionId",
-        "endPositionId",
-        "flux",
-        "bestWilsonPlotPath",
-        "totalExposedDose",
-        "nominalMagnification",
-        "nominalDefocus",
-        "imageSizeX",
-        "imageSizeY",
-        "pixelSizeOnImage",
-        "phasePlate",
-        "dataCollectionPlanId",
-    ]
 }
 
 GRID_INFO_COLUMN_MAP = {
@@ -625,33 +518,3 @@ def generate_scan_data_infos(
     else:
         scan_data_infos = [xy_scan_data_info]
     return scan_data_infos
-
-
-def compare_actual_and_expected(
-    id, expected_values, fetch_datacollection_attribute, column_map: dict | None = None
-):
-    results = "\n"
-    for k, v in expected_values.items():
-        actual = fetch_datacollection_attribute(
-            id, column_map[k.lower()] if column_map else k
-        )
-        if isinstance(actual, Decimal):
-            actual = float(actual)
-        if isinstance(v, float):
-            actual_v = actual == pytest.approx(v)
-        else:
-            actual_v = actual == v
-        if not actual_v:
-            results += f"expected {k} {v} == {actual}\n"
-    assert results == "\n", results
-
-
-def compare_comment(
-    fetch_datacollection_attribute, data_collection_id, expected_comment
-):
-    actual_comment = fetch_datacollection_attribute(
-        data_collection_id, DATA_COLLECTION_COLUMN_MAP["comments"]
-    )
-    match = re.search(" Zocalo processing took", actual_comment)
-    truncated_comment = actual_comment[: match.start()] if match else actual_comment
-    assert truncated_comment == expected_comment
